% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{.parameter_definitions}
\alias{.parameter_definitions}
\title{Defines some commonly used parameters}
\usage{
.parameter_definitions()
}
\arguments{
\item{synapse_id}{The Synapse ID of the file.}
}
\description{
Defines some commonly used parameters
}
