% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluationQuery}
\alias{evaluationQuery}
\title{Query an Evaluation object}
\usage{
evaluationQuery(evaluation_id, as_data_frame = TRUE)
}
\arguments{
\item{evaluation_id}{The ID of the evaluation object.}

\item{as_data_frame}{Return results as a dataframe. If FALSE, returns the
object returned by the API service `GET /evaluation/submission/query`.}
}
\value{
See the parameter description for `as_data_frame`.
}
\description{
Fetches all key/value pairs associated with the submissions objects in an
evaluation. Includes both annotations and default Evaluation values such as
name, submitterId, objectId, etc.
}
