% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{synGetFile}
\alias{synGetFile}
\title{Gets a file on Synapse.}
\usage{
synGetFile(synapse_id, reader = readr::read_csv, ...)
}
\arguments{
\item{synapse_id}{The Synapse ID of the file.}

\item{reader}{A function which accepts a file path as its first argument.
Used to read the file into a data.frame, tibble, etc. Default is to assume
the file is a CSV file and to return it as a tibble.}

\item{...}{Additional arguments to be passed to synapser::synGet}
}
\description{
Gets a file on Synapse.
}
