% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{synGetTable}
\alias{synGetTable}
\title{Get a table on Synapse}
\usage{
synGetTable(
  synapse_id,
  columns = list(),
  conditionals = list(),
  as_query_str = FALSE,
  as_data_frame = TRUE
)
}
\arguments{
\item{synapse_id}{The Synapse ID of the table.}

\item{columns}{A character vector of columns to download. Defaults to all columns.}

\item{conditionals}{A character vector of SQL conditionals.}

\item{as_query_str}{If TRUE, returns the query string without executing the query.
Defaults to FALSE.}

\item{as_data_frame}{If TRUE, returns the table as a dataframe. Defaults to TRUE.}
}
\description{
Downloads a table on Synapse. Either as a dataframe, a query object, or returns
the query string by itself.
}
