% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{synGetTableFiles}
\alias{synGetTableFiles}
\title{Get a table and its files}
\usage{
synGetTableFiles(
  synapse_id,
  file_columns,
  limit = NULL,
  offset = NULL,
  conditionals = list()
)
}
\arguments{
\item{synapse_id}{The Synapse ID of the table.}

\item{file_columns}{A character vector of columns which contain file
handles that need downloading.}

\item{limit}{Upper limit on the number of rows returned.}

\item{offset}{Number of rows to offset in the SQL query.}

\item{conditionals}{A character vector of SQL conditionals.}
}
\description{
Downloads a table on Synapse as a dataframe and adds columns containing
local filepaths for the specified columns.
}
